Apply FINDINGS F1+F2 (u128 casts) + F3 calibration constraint
Founder-approved fixes from the final findings review. The two u128
casts are the ONLY logic changes to lib.rs since the handoff (besides
declare_id), both lossless in the entire normal operating range:

- sell: tax = gross_out * tax_bps now computed in u128 (F1). The old
  u64 product panicked above gross ~ u64::MAX/tax_bps (7.4M SOL at the
  25% tax cap).
- resolve: fee = pot_total * protocol_fee_bps now computed in u128
  (F2). The old u64 product bricked the unchunkable resolve above pot
  ~ 18.4M SOL at the 10% fee cap - the one instruction that must never
  brick. Each cast carries an invariant comment.

Tests flipped accordingly (anti-regression in the new direction):
- F1 test: a sell whose gross x tax_bps exceeds u64 (9M SOL whale,
  25% tax) now succeeds with the exact mirrored tax routed to the
  prize vault.
- F2 test: resolve on a ~20M SOL pot succeeds, takes the exact 10% fee
  and freezes exact snapshots.

F3 (rounding profit corner above 1 lamport/raw-unit) stays a
calibration constraint, NOT a curve patch: check_f3() added to
tools/calibrate.py, and its VTOK0 default raised from 1e9 to 1e15 raw
units - the old default itself violated the constraint (4000 lamports
per unit at peak); token scale cancels out of every simulated
multiple, so economic outputs are unchanged. FINDINGS.md updated with
the decisions (O2 overflow-checks ratified).

Verified: anchor build green, 34/34 tests passing.
Devnet redeploy of the fixed binary pending faucet funds (upgrade
buffer needs ~2.9 SOL, wallet at 1.7).

+++ programs/ephemere/src/lib.rs
@@ -281,7 +281,9 @@ pub mod ephemere {
         require!(gross_out <= m.real_reserve, EphemereError::InsufficientReserve);
 
         let tax_bps = ev.sell_tax_bps[ev.current_round as usize] as u64;
-        let tax = gross_out * tax_bps / BPS_DENOM;
+        // u128: the bps product of a u64 amount can exceed u64 — the result
+        // is ≤ gross_out, so the cast back is lossless (FINDINGS F1).
+        let tax = ((gross_out as u128) * (tax_bps as u128) / BPS_DENOM as u128) as u64;
         let net_out = gross_out - tax;
         require!(net_out >= min_lamports_out, EphemereError::Slippage);
 
@@ -421,7 +423,11 @@ pub mod ephemere {
             .prize_vault
             .lamports()
             .saturating_sub(rent_floor);
-        let fee = pot_total * ev.protocol_fee_bps as u64 / BPS_DENOM;
+        // u128: resolve is the one instruction that must never brick — the
+        // bps product of a u64 pot can exceed u64 (FINDINGS F2); the result
+        // is ≤ pot_total, so the cast back is lossless.
+        let fee =
+            ((pot_total as u128) * (ev.protocol_fee_bps as u128) / BPS_DENOM as u128) as u64;
         if fee > 0 {
             debit(&ctx.accounts.prize_vault, fee)?;
             credit(&ctx.accounts.treasury, fee)?;
